requirements.txt
matplotlib
pandas
streamlit
scikit-learn
python-dotenv